backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
requests==2.32.3
pdfminer.six==20231228
numpy==1.26.4
scipy==1.13.0
reportlab==4.2.0
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0